+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
  - More `gen.h` macros: `ML99_GEN_SYM`, `ML99_TRAILING_SEMICOLON`, `ML99_semicoloned`, `ML99_assign`.
- - More statement chaining macros: `ML99_CHAIN_EXPR_STMT`, `ML99_INTRODUCE_NON_NULL_PTR_TO_STMT`.
+ - More statement chaining macros: `ML99_CHAIN_EXPR_STMT`, `ML99_CHAIN_EXPR_STMT_AFTER`, `ML99_INTRODUCE_NON_NULL_PTR_TO_STMT`.
  - The ability to specify several variables in `ML99_INTRODUCE_VAR_TO_STMT`.
  - `ML99_DETECTOR` handling all the C11 keywords and the underscore character (`_`).
  - `ML99_uncomma`